requirements.txt
Flask~=3.1.0
firebase-admin~=6.7.0
pyrebase4==4.7.1
Flask-WTF==1.1.1
python-dotenv~=1.1.0
wtforms~=3.2.1
Flask-Login~=0.6.3
werkzeug~=3.1.3
protobuf~=5.29.4
google-api-python-client~=2.166.0
dotenv~=0.9.9